RUBY-1279 Move shared modules to shared directory
