+++ lib/mongo/operation/shared/object_id_generator.rb
