+++ lib/mongo/operation/shared/specifiable.rb
